% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultipleBreakpoints.R
\name{plot.mBP}
\alias{plot.mBP}
\title{Plotting the Results of the \link{multiple_breakpoints} function}
\usage{
\method{plot}{mBP}(x, ask=TRUE, ...)
}
\arguments{
\item{x}{The result of a call to \link{multiple_breakpoints}}

\item{ask}{logical value. If \code{TRUE} (and the R session is interactive)
the user is asked for input, before a new figure is drawn
(see \code{\link{devAskNewPage}}).}

\item{...}{Further arguments are currently ignored. Only for compatibility
with generic functions.}
}
\description{
Plotting the empirical processes, the success probabilities and
breakpoints estimated by the \link{multiple_breakpoints} function
}
\examples{
mBP <- multiple_breakpoints(matrix(c(sort(rnorm(2000)),
                            rbinom(1000, 1, 0.2),
                            rbinom(1000, 1, 0.6)), ncol = 2))
plot(mBP)
}
\references{
\insertRef{froelich}{MultipleBreakpoints}
}
\author{
Nicolas Froelich
}
